requirements.txt
black==24.8.0
django-cors-headers==4.4.0
djoser==2.2.3
drf-spectacular==0.27.2
flake8==7.1.1
isort==5.13.2
pip-chill==1.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
